Move email to a single file
